requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.2.3
docutils==0.15.2
extras==1.0.0
idna==2.8
lockfile==0.12.2
multidict==4.5.2
pbr==5.4.1
pydantic==0.31
python-daemon==2.2.3
six==1.12.0
websockets==6.0
yarl==1.3.0
